requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.0
colorama==0.4.4
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
joblib==1.1.0
lxml==4.8.0
multitasking==0.0.10
nltk==3.7
numpy==1.22.3
oauthlib==3.2.0
pandas==1.4.1
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.1
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
textblob==0.17.1
tqdm==4.63.1
tweepy==4.8.0
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
yfinance==0.1.70
